backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
supabase>=0.7.1
pydantic>=1.8.0
python-dotenv>=0.19.0 